Fix conversions signed/unsigned
Inline functions are compiled with options of the client of the library.
Setting explicit conversion disable unexpected warnings.

+++ ntirpc/rpc/xdr.h
@@ -503,7 +503,7 @@ xdr_getint32(XDR *xdrs, int32_t *ip)
 static inline bool
 xdr_putint32(XDR *xdrs, int32_t v)
 {
-	return xdr_putuint32(xdrs, v);
+	return xdr_putuint32(xdrs, (uint32_t)v);
 }
 
 #define XDR_GETINT32(xdrs, int32p) xdr_getint32(xdrs, int32p)
@@ -553,7 +553,7 @@ xdr_getint16(XDR *xdrs, int16_t *ip)
 static inline bool
 xdr_putint16(XDR *xdrs, int32_t int16v)
 {
-	return xdr_putuint16(xdrs, int16v);
+	return xdr_putuint16(xdrs, (uint16_t)int16v);
 }
 
 #define XDR_GETINT16(xdrs, int16p) xdr_getint16(xdrs, int16p)
@@ -603,7 +603,7 @@ xdr_getint8(XDR *xdrs, int8_t *ip)
 static inline bool
 xdr_putint8(XDR *xdrs, int32_t int8v)
 {
-	return xdr_putuint8(xdrs, int8v);
+	return xdr_putuint8(xdrs, (uint8_t)int8v);
 }
 
 #define XDR_GETINT8(xdrs, int8p) xdr_getint8(xdrs, int8p)
